--- a/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/BitConverterCompat.cs
+++ b/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/BitConverterCompat.cs
@@ -1,0 +1,18 @@
+// See the LICENSE file in the project root for more information.
+
+using System;
+
+namespace Microsoft.Data.SqlClient
+{
+    internal static class BitConverterCompat
+    {
+        public static unsafe int SingleToInt32Bits(float value) 
+        {
+           return *(int*)(&value);
+        }
+        public static unsafe float Int32BitsToSingle(int value) 
+        {
+            return *(float*)(&value);
+        }       
+    }
+}