--- a/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParserStateObject.netcore.cs
+++ b/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParserStateObject.netcore.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Diagnostics;
+using System.Buffers.Binary;
 using System.Runtime.InteropServices;
 using System.Security;
 using System.Threading;
@@ -527,7 +528,7 @@ namespace Microsoft.Data.SqlClient
                     Debug.Assert(_bTmpRead + bytesRead == 8, "TryReadByteArray returned true without reading all data required");
                     _bTmpRead = 0;
                     AssertValidState();
-                    value = BitConverter.ToInt64(_bTmp, 0);
+                    value = BinaryPrimitives.ReadInt64LittleEndian(_bTmp);
                     return true;
                 }
             }
@@ -536,7 +537,7 @@ namespace Microsoft.Data.SqlClient
                 // The entire long is in the packet and in the buffer, so just return it
                 // and take care of the counters.
 
-                value = BitConverter.ToInt64(_inBuff, _inBytesUsed);
+                value = BinaryPrimitives.ReadInt64LittleEndian(_inBuff.AsSpan(_inBytesUsed));
 
                 _inBytesUsed += 8;
                 _inBytesPacket -= 8;
@@ -605,7 +606,7 @@ namespace Microsoft.Data.SqlClient
                     Debug.Assert(_bTmpRead + bytesRead == 4, "TryReadByteArray returned true without reading all data required");
                     _bTmpRead = 0;
                     AssertValidState();
-                    value = BitConverter.ToUInt32(_bTmp, 0);
+                    value = BinaryPrimitives.ReadUInt32LittleEndian(_bTmp);
                     return true;
                 }
             }
@@ -614,7 +615,7 @@ namespace Microsoft.Data.SqlClient
                 // The entire int is in the packet and in the buffer, so just return it
                 // and take care of the counters.
 
-                value = BitConverter.ToUInt32(_inBuff, _inBytesUsed);
+                value = BinaryPrimitives.ReadUInt32LittleEndian(_inBuff.AsSpan(_inBytesUsed));
 
                 _inBytesUsed += 4;
                 _inBytesPacket -= 4;
@@ -639,7 +640,7 @@ namespace Microsoft.Data.SqlClient
                 }
 
                 AssertValidState();
-                value = BitConverter.ToSingle(_bTmp, 0);
+                value = BitConverterCompat.Int32BitsToSingle(BinaryPrimitives.ReadInt32LittleEndian(_bTmp));
                 return true;
             }
             else
@@ -647,7 +648,7 @@ namespace Microsoft.Data.SqlClient
                 // The entire float is in the packet and in the buffer, so just return it
                 // and take care of the counters.
 
-                value = BitConverter.ToSingle(_inBuff, _inBytesUsed);
+                value = BitConverterCompat.Int32BitsToSingle(BinaryPrimitives.ReadInt32LittleEndian(_inBuff.AsSpan(_inBytesUsed)));
 
                 _inBytesUsed += 4;
                 _inBytesPacket -= 4;
@@ -672,7 +673,7 @@ namespace Microsoft.Data.SqlClient
                 }
 
                 AssertValidState();
-                value = BitConverter.ToDouble(_bTmp, 0);
+                value = BitConverter.Int64BitsToDouble(BinaryPrimitives.ReadInt64LittleEndian(_bTmp));
                 return true;
             }
             else
@@ -680,7 +681,7 @@ namespace Microsoft.Data.SqlClient
                 // The entire double is in the packet and in the buffer, so just return it
                 // and take care of the counters.
 
-                value = BitConverter.ToDouble(_inBuff, _inBytesUsed);
+                value = BitConverter.Int64BitsToDouble(BinaryPrimitives.ReadInt64LittleEndian(_inBuff.AsSpan(_inBytesUsed)));
 
                 _inBytesUsed += 8;
                 _inBytesPacket -= 8;
@@ -1793,6 +1794,15 @@ namespace Microsoft.Data.SqlClient
                             str = Marshal.SecureStringToBSTR(_securePasswords[i]);
                             byte[] data = new byte[_securePasswords[i].Length * 2];
                             Marshal.Copy(str, data, 0, _securePasswords[i].Length * 2);
+                            if (!BitConverter.IsLittleEndian)
+                            {
+                                Span<byte> span = data.AsSpan();
+                                for (int ii = 0; ii < _securePasswords[i].Length * 2; ii += 2)
+                                {
+                                    short value = BinaryPrimitives.ReadInt16LittleEndian(span.Slice(ii));
+                                    BinaryPrimitives.WriteInt16BigEndian(span.Slice(ii), value);
+                                }
+                            }
                             TdsParserStaticMethods.ObfuscatePassword(data);
                             data.CopyTo(_outBuff, _securePasswordOffsetsInBuffer[i]);
                         }
@@ -2294,7 +2304,8 @@ namespace Microsoft.Data.SqlClient
                 // So we need to avoid this check prior to login completing
                 state == TdsParserState.OpenLoggedIn
                     && !_bulkCopyOpperationInProgress // ignore the condition checking for bulk copy
-                    && _outBytesUsed == (_outputHeaderLen + BitConverter.ToInt32(_outBuff, _outputHeaderLen))
+                    && _outBytesUsed == (_outputHeaderLen +
+                    BinaryPrimitives.ReadInt32LittleEndian(_outBuff.AsSpan(_outputHeaderLen)))
                     && _outputPacketCount == 0
                     || _outBytesUsed == _outputHeaderLen
                     && _outputPacketCount == 0)