--- a/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParserStateObject.netcore.cs
+++ b/src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParserStateObject.netcore.cs
@@ -279,7 +279,6 @@ namespace Microsoft.Data.SqlClient
             _cancellationOwner.Target = cancellationOwner;
         }
 
-
         /////////////////////////////////////////
         // Value Skip Logic                    //
         /////////////////////////////////////////