The Microsoft SqlClient driver crashes on the s390x. This error stems mostly from the use of BitConverter function used in method definitions across the code and improper byte ordering for big endian systems.

The BitConverter class does conversions according to the endianess of the system it is operating on which causes incorrect reads of little-endian packets sent by the server.