@JRahnama 2023-09-28
@saitama951 failures are not related to your changes. Somehow starting SqlBrowser task fails on the pipelines. We will look into them.