@Wraith2 2023-09-29
I'm a little concerned about the number of preprocessor regions. If there's any way to extract common patterns into functions so we have fewer ifdefs it'd be nice. Otherwise without the ability to test i'll go with your changes as long as you've got the tests working.

Out of interest are you using managed networking? or is there a big endian native sni.dll ?