@saitama951 2023-10-02
> I'm a little concerned about the number of preprocessor regions. If there's any way to extract common patterns into functions so we have fewer ifdefs it'd be nice. Otherwise without the ability to test i'll go with your changes as long as you've got the tests working.
> 
> Out of interest are you using managed networking? or is there a big endian native sni.dll ?

I agree with you  the if defs are unnecessarily populating the code, I do see the csproj includes a package System.Memory which supports BinaryPrimitives for all the targets but some functions are not supported with netstandard2.0 so I have added a file with the compatibility support, let me know your thoughts on this.
to answer your second question it uses the managed networking dll.