@DavoudEshtehari 2023-11-03
My concern is how this change is going to impact the performance by impacting the hot paths. Does anybody have a perf test to compare it with before the change?