@saitama951 2023-11-10
@DavoudEshtehari 
I did manage to run the performance tests and here are the results
note:- the base= without the patch and diff= with the patch
```
summary:
better: 1, geomean: 3.491
worse: 42, geomean: 1.254
total diff: 43

| Slower                                                                                        | diff/base | Base Median (ns) | Diff Median (ns) | Modality  |
| --------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------- |
| Microsoft.Data.SqlClient.PerformanceTests.SqlCommandRunner.ExecuteScalar                      |      1.52 |      81632944.25 |     123795769.75 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.SqlCommandRunner.ExecuteXmlReader                   |      1.41 |     362968055.00 |     511413731.50 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.DateTimeOffset                 |      1.40 |      17359631.75 |      24297528.00 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.SmallDateTime                  |      1.36 |      19801378.00 |      27024967.50 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Decimal                        |      1.35 |      18220311.50 |      24638132.50 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.VarBinary                      |      1.35 |      24514042.75 |      33120146.50 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.NChar                          |      1.35 |      18399459.50 |      24820209.00 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.NVarChar                       |      1.34 |      25595037.50 |      34186686.75 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Datetime                       |      1.33 |      18858414.50 |      25097611.75 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Xml                            |      1.33 |      30359710.00 |      40349804.00 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Char                           |      1.31 |      18560508.00 |      24309393.25 |           |
| Microsoft.Data.SqlClient.PerformanceTests.SqlCommandRunner.ExecuteNonQuery                    |      1.31 |      97676723.75 |     127612822.00 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Datetime2                      |      1.29 |      19728887.00 |      25509330.50 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Binary                         |      1.29 |      18213765.00 |      23410165.50 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.NVarCharAsync             |      1.28 |      30038761.25 |      38505020.00 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Real                           |      1.27 |      18352333.00 |      23353771.75 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Date                           |      1.26 |      18424424.50 |      23286289.25 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Float                          |      1.26 |      18732205.75 |      23654323.75 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.SmallMoney                     |      1.25 |      19230988.00 |      24130540.50 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Numeric                        |      1.24 |      19406590.00 |      24090994.00 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.VarCharAsync              |      1.24 |      29144578.50 |      36085790.00 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.VarChar                        |      1.24 |      25736796.75 |      31856117.50 | multimodal|
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.NumericAsync              |      1.24 |      24404761.50 |      30154729.50 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.BitAsync                  |      1.22 |      23455534.50 |      28669823.25 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.CharAsync                 |      1.21 |      24139974.50 |      29258859.50 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.DateTimeOffsetAsync       |      1.21 |      23872583.00 |      28934241.75 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Time                           |      1.21 |      20120527.00 |      24290295.25 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.IntAsync                  |      1.20 |      24145428.25 |      29087973.25 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.DecimalAsync              |      1.20 |      25246572.00 |      30361482.00 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.Datetime2Async            |      1.20 |      25484344.75 |      30597239.50 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.DatetimeAsync             |      1.20 |      25987078.75 |      31198285.00 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.VarBinaryAsync            |      1.20 |      30502607.50 |      36583319.25 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.SmallIntAsync             |      1.20 |      24918030.75 |      29813608.25 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.SqlBulkCopyRunner.BulkCopy_IDataReader(Columns: 25) |      1.19 |     632961417.50 |     753515670.50 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.SmallDateTimeAsync        |      1.18 |      26242335.25 |      30973506.75 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.TimeAsync                 |      1.18 |      25614011.75 |      30228362.25 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.SmallMoneyAsync           |      1.18 |      25276536.00 |      29796326.75 | multimodal|
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.BigIntAsync               |      1.16 |      25780506.25 |      30006454.00 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.UniqueIdentifierAsync     |      1.16 |      24704125.75 |      28671626.50 |           |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.NCharAsync                |      1.16 |      24424800.50 |      28304558.75 | bimodal   |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.XmlAsync                  |      1.15 |      34468910.50 |      39579307.75 | several?  |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderAsyncRunner.BinaryAsync               |      1.14 |      25671757.00 |      29162259.00 |           |

| Faster                                                               | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| Microsoft.Data.SqlClient.PerformanceTests.DataTypeReaderRunner.Money |      3.49 |      84750981.50 |      24273529.00 | bimodal |
```

I did manage to run the perf profiler to look out for hotspots in the code for the first couple of benchmarks.
what I could find is that most of the hotspots are in `BinaryPrimitives` i.e `ReadUInt32`

I'm looking into the other regressions as well 
NOTE :- There might be false positives among these, since these are running on a VM  (both server and client).